refac: package structure for watcher testing
To test the watcher realistically the tested modules in "./watchertest"
should import the pyunit package.  This revealed with the previous
package structure that one had to import "from pyunit import pyunit".
To avoid stuttering the pyunit module was renamed to exec and is now
only about running test suites while __init__.py was added and provides
the function "run_tests", the module "reporting" as well as the types
"Config" and "T".  Which as of now are the public API of pyunit.  With
an appropriate appropriate "context.py" in "./watchertest" a test module
can now say "from context import pyunit" and can use its features like
"pyunit.run_tests(Suite)" etc.