refac: we can't export on package level ...
... without adding the package to the import path which contradicts the
name spacing purpose of a package.  We always have something todo like
    "from pkg import module"
and if we want to use in module m1 from pkg a feature f in module m2 we
need to do
    "from m2 import f"
(we ignore relative imports) the way of Go "pkg.feature" seems in spite
of flat is better than nested not possible.  To avoid from pyunit import
pyunit the module is now called testing while the testing-module is now
called assert_.  Hence the typical usage of pyunit is now:

from pyunit import testing

class MySuite:

    def a_test(self, t: testing.T)

if __name__ == '__main__':
    testing.run(MySuite, testing.Config(...))

which reads well, I think.