refac(packaging III): minimized user available api
Because of python's lack of a privacy concept all modules in the pyunit
package were available for import and all their objects inside.  Hence
all modules except for the testing module were moved inside an _internal
sub-package.  All objects imported into testing which are not meant to
be imported by a user of pyunit were dunder renamed.  It turned out that
we can't ignore relative import if we want IDE (vscode) support for
imports of _internal modules in testing, i.e. writing in testing
    from pyunit._internal import reporting
 works practically but vscode tells
    us that it doesn't find pyunit since
it is not smart enough to add the package path to the code analysis.
Doing that through configuration I couldn't make work.  On the other
hand
    from ._internal import reporting
works.  During my investigations I learned that having a "tests"
sub-package where the test modules live seems to be a convention as well
as naming such modules test_*.py.  I prefer  to have my tests next to
the code which is tested hence I chose to name them *_test.py and put
them in the same package as the production code.  This is not a problem
for the tooling so far but it adds additional complexity to the watcher.
It can now not only watch *_test.py files in the package directory but
it should also check for a tests and for an _internal package in the
directory of the package.  If there are sub-packages with identical
module names and a central tests package it is not possible to match the
test modules to the production modules only by their file names, i.e. we
would have to parse the test modules to figure which production modules
they import to know which tests to rerun if a production module is
changed.  Hence actually a watcher should find all test_*.py and
*_test.py modules in the watched directory and in a potentially parent
tests sub-package and in all subordinate packages and parse their
imports to map production code to test code.