refac(packaging IV): rm relative package import
Relative import in testing lead to import problems for dynamic module
loading in the watcher when only the parent of the package path is
dynamically added to the search path.  Also in _internal/testcontext an
adding of the package-path itself was necessary to run the tests which
already mentioned in a previous commit contradicts the package name
spacing.  But I found the vscode setting:
    "python.analysis.extraPaths": ["path to package"]
does the trick.  I.e. we are
    - finally rid of relative imports,
    - the usage of the package API for a user functions,
    - all tests run,
    - the watcher can with dynamic imports figure out the production
      modules which are imported by a test module
    - testing.py's absolute imports are found by Pylance.
Hallelujah!