refac: for multiprocessing and tui
*run_test_modules* became *run_test_module* and executes only one test
module and collects its output.  Hence we can map *test_modules_to_run*
on *run_test_module* in a multiprocessing worker pool.

Further more a tui module was added.  This indirection encapsulates the
terminal access and is parameterized with an text-io instance.  The
later allows for tests to redirect the output to an in-memory buffer
while the former makes it possible to add/exchange any tui-library
without changing the watcher code.