refac: production dependency calculations parallel
Since we already have a worker pool for executing test modules in
parallel it is also leveraged to calculate the production dependencies
of all watched test modules in parallel.  Since production dependencies
are only calculated in case a production module has been updated the
need for calculating production dependencies is pushed down to the
Snapshot type which calculates the test-modules to rerun between two run
cycles.  Hence the watching Dir type became an other optional argument
which is a function providing production dependencies for a set of test
modules.  This function defaults to the static TM.dependencies
calculating them in series while "main" may override this default with a
function calculating them in parallel.